Update buildinfo.json, index.js, and timestamp
This commit updates the buildinfo.json, index.js, and timestamp files. The changes include updating the build timestamp and version number. The new build timestamp is 1727712448 and the new version number is "39288b8". This commit also removes the unused sync.yml file and adds the qweather.json file to the plugins directory.